Remove and combine deep q environment
